Changed AutoMockPropertiesCommand to correctly handle null specimens
