+++ Src/AutoMoq/AutoMockPropertiesCommand.cs
@@ -23,7 +23,6 @@ namespace Ploeh.AutoFixture.AutoMoq
         /// <param name="context">The context that is used to create anonymous values.</param>
         public void Execute(object specimen, ISpecimenContext context)
         {
-            if (specimen == null) throw new ArgumentNullException("specimen");
             if (context == null) throw new ArgumentNullException("context");
 
             var mock = specimen as Mock;
